+++ R/phrase.R
@@ -9,10 +9,8 @@
 #' @examples
 #' x <- c('pop', 'pop_2020_est', 'pop_white_2020', 'pop_black_2020')
 #' rem_phrase(x, '_2020')
-rem_phrase <- function(x, phrase){
-  x <- stringr::str_remove(x, phrase)
-  x[nchar(x) == 0] <- x[nchar(x) == 0]
-  x
+rem_phrase <- function(x, phrase) {
+  stringr::str_remove_all(x, stringr::fixed(phrase))
 }
 
 #' Replace Phrase
@@ -27,6 +25,6 @@ rem_phrase <- function(x, phrase){
 #' @examples
 #' x <- c('pop', 'pop_2020_est', 'pop_white_2020', 'pop_black_2020')
 #' repl_phrase(x, '_2020', '_20')
-repl_phrase <- function(x, phrase, repl){
-  stringr::str_replace_all(x, phrase, repl)
+repl_phrase <- function(x, phrase, repl) {
+  stringr::str_replace_all(x, stringr::fixed(phrase), repl)
 }