+++ R/sort.R
@@ -53,15 +53,16 @@ sort_phrase <- function(x, loc = 2) {
 #' sort_suff(x)
 sort_suff <- function(x) {
   to_end <- x[!grepl('_', x)]
-  pieces <- strsplit(x[grep('_', x)], '_')
+  x_under <- x[grep('_', x)]
+  pieces <- strsplit(x_under, '_')
   lens <- lengths(pieces)
   suffs <- sapply(seq_along(pieces), FUN = function(i) pieces[[i]][lens[i]])
   suff_ord <- sort(unique(suffs))
   suff_grp <- match(suffs, suff_ord)
   grp_ct <- c(0, cumsum(unname(table(suff_grp))))
-  new_ord <- vector('integer', length(x))
+  new_ord <- vector('integer', length(pieces))
   for (i in seq_along(unique(suff_ord))) {
     new_ord[suff_grp == i] <- seq(grp_ct[i] + 1L, grp_ct[i + 1L])
   }
-  c(x[match(seq_len(length(x)), new_ord, nomatch = 0)], to_end)
+  c(x_under[match(seq_len(length(pieces)), new_ord)], to_end)
 }