+++ R/suffix.R
@@ -9,7 +9,7 @@
 #' @examples
 #' x <- c('pop', 'pop_2020_est', 'pop_white_2020', 'pop_black_2020')
 #' add_suff(x, '_cen')
-add_suff <- function(x, suff){
+add_suff <- function(x, suff) {
   paste0(x, suff)
 }
 
@@ -24,7 +24,7 @@ add_suff <- function(x, suff){
 #' @examples
 #' x <- c('pop', 'pop_2020_est', 'pop_white_2020', 'pop_black_2020')
 #' rem_suff(x, '_2020')
-rem_suff <- function(x, suff){
+rem_suff <- function(x, suff) {
   nc <- stringr::str_length(suff)
   w <- stringr::str_sub(x, start = -nc) == suff
   x[w] <- stringr::str_sub(x[w], end = -(nc + 1))
@@ -43,7 +43,7 @@ rem_suff <- function(x, suff){
 #' @examples
 #' x <- c('pop', 'pop_2020_est', 'pop_white_2020', 'pop_black_2020')
 #' repl_suff(x, '_2020', '_20')
-repl_suff <- function(x, suff, repl){
+repl_suff <- function(x, suff, repl) {
   nc <- stringr::str_length(suff)
   w <- stringr::str_sub(x, start = -nc) == suff
   x[w] <- stringr::str_sub(x[w], end = -(nc + 1))