+++ tests/testthat.R
@@ -1,4 +1,4 @@
 library(testthat)
 library(name)
 
-test_check("name")
+test_check('name')